Refactoring iotdb inheritance structure
