+++ configuration/conf/config.properties
@@ -4,10 +4,10 @@
 
 ################### 被测数据库配置 #######################
 # 被测试的数据库，目前的格式为{name}{-version}{-insert mode}(注意-号)其全部参考值参见README文件，注意版本需要匹配
-# DB_SWITCH=IoTDB-130-SESSION_BY_TABLET
+ DB_SWITCH=IoTDB-130-SESSION_BY_TABLET
 
 # 是否使用 iotdb 表模型。iotdb-140 支持树模型与表模型。默认值为 false。
-# IoTDB_ENABLE_TABLE=false
+ IoTDB_ENABLE_TABLE=false
 
 # 主机列表，如果有多个主机则使用英文逗号进行分割
 # HOST=127.0.0.1
@@ -31,7 +31,7 @@
 
 ############### 数据库初始化：清理数据 ####################
 # 测试启动前是否清理数据
-# IS_DELETE_DATA=true
+ IS_DELETE_DATA=false
 
 # IS_DELETE_DATA = true 时生效，单位为ms，如果清理数据，则需要等待INIT_WAIT_TIME时间待数据库数据清理完成，当数据库为KairosDB, InfluxDb, OpenTSDB, TimescaleDB时必须
 # INIT_WAIT_TIME=1000
@@ -45,14 +45,14 @@
 
 ################## Benchmark配置 #######################
 # 总操作次数：具体每种类型操作会按OPERATION_PROPORTION定义的比例划分，例如OPERATION_PROPORTION=1:1:1:1:1:1:1:1:1:1:1
-# LOOP=100
+ LOOP=1
 
 # 目前Benchmark支持多种运行模式，支持模式如下
 # testWithDefaultPath   常规测试模式，支持多种读和写操作的混合负载
 # generateDataMode      生成数据模式，生成Benchmark本身识别的数据
 # verificationWriteMode 单数据库正确性写入模式，需要配置 FILE_PATH 以及 DATA_SET
 # verificationQueryMode 单数据库正确性查询模式，需要配置 FILE_PATH 以及 DATA_SET
-# BENCHMARK_WORK_MODE=testWithDefaultPath
+ BENCHMARK_WORK_MODE=testWithDefaultPath
 
 # 对于数据写入或查询，限制最长耗时，设置为0表示无限制，单位为ms
 # 对数据写入，最大误差约为WRITE_OPERATION_TIMEOUT_MS
@@ -72,9 +72,6 @@
 
 # 如果 BENCHMARK_CLUSTER=true, 则为当前Benchmark的编号
 # BENCHMARK_INDEX=0
-
-# 在集群模式下：是否所有数据库都可见，如果可见则可以向其他数据库节点发送数据
-# IS_ALL_NODES_VISIBLE=false
 
 ################ Benchmark：双写模式 ####################
 # 双写模式仅支持不同数据库之间进行比较，不支持同一个数据库不同版本进行双写
@@ -116,16 +113,16 @@
 
 ############## 数据：设备、传感器、客户端 ##################
 # 设备总数
-# DEVICE_NUMBER=6000
+ DEVICE_NUMBER=4
 
 # 实际写入设备比例，(0, 1]
 # REAL_INSERT_RATE=1.0
 
 # 每个设备的传感器总数
-# SENSOR_NUMBER=200
+ SENSOR_NUMBER=10
 
 # 各个传感器时间戳是否对齐
-# IS_SENSOR_TS_ALIGNMENT=true
+ IS_SENSOR_TS_ALIGNMENT=true
 
 # 当传感器不对齐时，每个设备的传感器的保留比例，(0, 1]
 # TS_ALIGNMENT_RATIO=1.0
@@ -140,7 +137,7 @@
 # SENSOR_NAME_PREFIX=s_
 
 # 标签名数量
-# TAG_NUMBER=0
+ TAG_NUMBER=1
 
 # 标签名前缀
 # TAG_KEY_PREFIX=tag_
@@ -149,13 +146,13 @@
 # TAG_VALUE_PREFIX=value_
 
 # 每个标签值的取值种数，使用逗号分开，设置个数应与标签名数量相等。例如TAG_NUMBER=2，TAG_VALUE_CARDINALITY=50,10，说明第一个标签名可能有50种值，第二个标签名可能有10种值。d_0的中间路径将为value_0.value_0.d_0。该配置不影响总点数。
-# TAG_VALUE_CARDINALITY=
+ TAG_VALUE_CARDINALITY=4
 
 # 是否将设备绑定给客户端，如果绑定，则客户端数小于等于设备数，否则可以大于
 # IS_CLIENT_BIND=true
 
 # 客户端总数
-# CLIENT_NUMBER=20
+ CLIENT_NUMBER=2
 
 ############## 被测系统为IoTDB时扩展参数 ##################
 # 是否使用thrift压缩，需要在iotdb的配置文件iotdb-datanode.properties中设置dn_rpc_thrift_compression_enable=true
@@ -234,7 +231,7 @@
 # DOUBLE_LENGTH=2
 
 # 插入数据的数据类型的比例，BOOLEAN:INT32:INT64:FLOAT:DOUBLE:TEXT:STRING:BLOB:TIMESTAMP:DATE
-# INSERT_DATATYPE_PROPORTION=1:1:1:1:1:1:0:0:0:0
+ INSERT_DATATYPE_PROPORTION=1:1:1:1:1:1:1:1:1:1
 
 ################ IoTDB相关元数据参数 #####################
 # 压缩方式 UNCOMPRESSED | SNAPPY | LZ4 (仅对IoTDB有效)
@@ -314,7 +311,7 @@
 # Q9  倒序范围查询（只限制起止时间）select v1... from data where time > ? and time < ? and device in ? order by time desc
 # Q10 倒序带值过滤的范围查询 select v1... from data where time > ? and time < ? and v1 > ? and device in ? order by time desc
 # Q11 分组聚合查询，倒序；目前仅支持iotdb、tdengine-3.0、influxdb v1
-# OPERATION_PROPORTION=1:0:0:0:0:0:0:0:0:0:0:0
+ OPERATION_PROPORTION=1:0:0:0:0:0:0:0:0:0:0:0
 
 # 最长等待写时间，单位毫秒，即如果整个写操作在指定时间内没有返回，则终止此操作
 # WRITE_OPERATION_TIMEOUT_MS=120000
