+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/tsdb/DBFactory.java
@@ -48,11 +48,7 @@ public class DBFactory {
         case DB_IOT_140_SESSION_BY_TABLET:
         case DB_IOT_140_SESSION_BY_RECORD:
         case DB_IOT_140_SESSION_BY_RECORDS:
-          if (config.isIS_ALL_NODES_VISIBLE()) {
-            dbClass = Constants.IOTDB140_ROUNDROBIN_SESSION_CLASS;
-          } else {
-            dbClass = Constants.IOTDB140_SESSION_CLASS;
-          }
+          dbClass = Constants.IOTDB140_SESSION_CLASS;
           break;
           // IoTDB 1.3
         case DB_IOT_130_REST:
@@ -64,11 +60,7 @@ public class DBFactory {
         case DB_IOT_130_SESSION_BY_TABLET:
         case DB_IOT_130_SESSION_BY_RECORD:
         case DB_IOT_130_SESSION_BY_RECORDS:
-          if (config.isIS_ALL_NODES_VISIBLE()) {
-            dbClass = Constants.IOTDB130_ROUNDROBIN_SESSION_CLASS;
-          } else {
-            dbClass = Constants.IOTDB130_SESSION_CLASS;
-          }
+          dbClass = Constants.IOTDB130_SESSION_CLASS;
           break;
           // IoTDB 1.1
         case DB_IOT_110_JDBC:
@@ -77,11 +69,11 @@ public class DBFactory {
         case DB_IOT_110_SESSION_BY_TABLET:
         case DB_IOT_110_SESSION_BY_RECORD:
         case DB_IOT_110_SESSION_BY_RECORDS:
-          if (config.isIS_ALL_NODES_VISIBLE()) {
-            dbClass = Constants.IOTDB110_ROUNDROBIN_SESSION_CLASS;
-          } else {
-            dbClass = Constants.IOTDB110_SESSION_CLASS;
-          }
+          //          if (config.isIS_ALL_NODES_VISIBLE()) {
+          //            dbClass = Constants.IOTDB110_ROUNDROBIN_SESSION_CLASS;
+          //          } else {
+          dbClass = Constants.IOTDB110_SESSION_CLASS;
+          //          }
           break;
           // IoTDB 1.0
         case DB_IOT_100_JDBC:
@@ -90,11 +82,7 @@ public class DBFactory {
         case DB_IOT_100_SESSION_BY_TABLET:
         case DB_IOT_100_SESSION_BY_RECORD:
         case DB_IOT_100_SESSION_BY_RECORDS:
-          if (config.isIS_ALL_NODES_VISIBLE()) {
-            dbClass = Constants.IOTDB100_ROUNDROBIN_SESSION_CLASS;
-          } else {
-            dbClass = Constants.IOTDB100_SESSION_CLASS;
-          }
+          dbClass = Constants.IOTDB100_SESSION_CLASS;
           break;
         case DB_INFLUX:
           dbClass = Constants.INFLUXDB_CLASS;
