+++ iotdb-1.3/src/main/java/cn/edu/tsinghua/iot/benchmark/iotdb130/IoTDB.java
@@ -161,41 +161,42 @@ public class IoTDB implements IDatabase {
       Map<Session, List<TimeseriesSchema>> sessionListMap = new HashMap<>();
       try {
         // open meta session
-        if (!config.isIS_ALL_NODES_VISIBLE()) {
-          Session metaSession =
-              new Session.Builder()
-                  .host(dbConfig.getHOST().get(0))
-                  .port(Integer.parseInt(dbConfig.getPORT().get(0)))
-                  .username(dbConfig.getUSERNAME())
-                  .password(dbConfig.getPASSWORD())
-                  .version(Version.V_1_0)
-                  .sqlDialect(dbConfig.getSQL_DIALECT())
-                  .build();
-          metaSession.open(config.isENABLE_THRIFT_COMPRESSION());
-          sessionListMap.put(metaSession, createTimeseries(schemaList));
-        } else {
-          int sessionNumber = dbConfig.getHOST().size();
-          List<Session> keys = new ArrayList<>();
-          for (int i = 0; i < sessionNumber; i++) {
-            Session metaSession =
-                new Session.Builder()
-                    .host(dbConfig.getHOST().get(i))
-                    .port(Integer.parseInt(dbConfig.getPORT().get(i)))
-                    .username(dbConfig.getUSERNAME())
-                    .password(dbConfig.getPASSWORD())
-                    .version(Version.V_1_0)
-                    .sqlDialect(dbConfig.getSQL_DIALECT())
-                    .build();
-            metaSession.open(config.isENABLE_THRIFT_COMPRESSION());
-            keys.add(metaSession);
-            sessionListMap.put(metaSession, new ArrayList<>());
-          }
-          for (int i = 0; i < schemaList.size(); i++) {
-            sessionListMap
-                .get(keys.get(i % sessionNumber))
-                .add(createTimeseries(schemaList.get(i)));
-          }
-        }
+        //        if (!config.isIS_ALL_NODES_VISIBLE()) {
+        Session metaSession =
+            new Session.Builder()
+                .host(dbConfig.getHOST().get(0))
+                .port(Integer.parseInt(dbConfig.getPORT().get(0)))
+                .username(dbConfig.getUSERNAME())
+                .password(dbConfig.getPASSWORD())
+                .version(Version.V_1_0)
+                .sqlDialect(dbConfig.getSQL_DIALECT())
+                .build();
+        metaSession.open(config.isENABLE_THRIFT_COMPRESSION());
+        sessionListMap.put(metaSession, createTimeseries(schemaList));
+        //        } else {
+        //
+        //          int sessionNumber = dbConfig.getHOST().size();
+        //          List<Session> keys = new ArrayList<>();
+        //          for (int i = 0; i < sessionNumber; i++) {
+        //            Session metaSession =
+        //                new Session.Builder()
+        //                    .host(dbConfig.getHOST().get(i))
+        //                    .port(Integer.parseInt(dbConfig.getPORT().get(i)))
+        //                    .username(dbConfig.getUSERNAME())
+        //                    .password(dbConfig.getPASSWORD())
+        //                    .version(Version.V_1_0)
+        //                    .sqlDialect(dbConfig.getSQL_DIALECT())
+        //                    .build();
+        //            metaSession.open(config.isENABLE_THRIFT_COMPRESSION());
+        //            keys.add(metaSession);
+        //            sessionListMap.put(metaSession, new ArrayList<>());
+        //          }
+        //          for (int i = 0; i < schemaList.size(); i++) {
+        //            sessionListMap
+        //                .get(keys.get(i % sessionNumber))
+        //                .add(createTimeseries(schemaList.get(i)));
+        //          }
+        //        }
 
         if (config.isTEMPLATE()
             && !config.isIoTDB_ENABLE_TABLE()
