+++ iotdb-1.3/src/main/java/cn/edu/tsinghua/iot/benchmark/iotdb130/IoTDBSession.java
@@ -25,171 +25,590 @@ import org.apache.iotdb.rpc.IoTDBConnectionException;
 import org.apache.iotdb.rpc.StatementExecutionException;
 import org.apache.iotdb.session.Session;
 
+import cn.edu.tsinghua.iot.benchmark.client.operation.Operation;
+import cn.edu.tsinghua.iot.benchmark.conf.Config;
+import cn.edu.tsinghua.iot.benchmark.conf.ConfigDescriptor;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.IBatch;
+import cn.edu.tsinghua.iot.benchmark.entity.DeviceSummary;
+import cn.edu.tsinghua.iot.benchmark.entity.Record;
+import cn.edu.tsinghua.iot.benchmark.entity.Sensor;
+import cn.edu.tsinghua.iot.benchmark.entity.enums.SensorType;
+import cn.edu.tsinghua.iot.benchmark.exception.OperationFailException;
+import cn.edu.tsinghua.iot.benchmark.measurement.Status;
+import cn.edu.tsinghua.iot.benchmark.schema.schemaImpl.DeviceSchema;
 import cn.edu.tsinghua.iot.benchmark.tsdb.DBConfig;
 import cn.edu.tsinghua.iot.benchmark.tsdb.TsdbException;
+import cn.edu.tsinghua.iot.benchmark.tsdb.enums.DBInsertMode;
+import cn.edu.tsinghua.iot.benchmark.workload.query.impl.DeviceQuery;
+import cn.edu.tsinghua.iot.benchmark.workload.query.impl.VerificationQuery;
 import org.apache.tsfile.enums.TSDataType;
+import org.apache.tsfile.file.metadata.enums.TSEncoding;
+import org.apache.tsfile.read.common.Field;
 import org.apache.tsfile.read.common.RowRecord;
+import org.apache.tsfile.utils.Binary;
+import org.apache.tsfile.utils.BytesUtils;
 import org.apache.tsfile.write.record.Tablet;
-import org.slf4j.LoggerFactory;
+import org.apache.tsfile.write.schema.IMeasurementSchema;
+import org.apache.tsfile.write.schema.MeasurementSchema;
+import org.slf4j.Logger;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
+import java.util.stream.Collectors;
 
-public class IoTDBSession extends IoTDBSessionBase {
-  private class BenchmarkSession implements IBenchmarkSession {
-    private final Session session;
+public class IoTDBSession extends IoTDB {
+  static Logger LOGGER;
 
-    public BenchmarkSession(Session session) {
-      this.session = session;
-    }
+  static final Config config = ConfigDescriptor.getInstance().getConfig();
+  //  IBenchmarkSession sessionWrapper;
 
-    @Override
-    public void open() throws IoTDBConnectionException {
-      session.open();
-    }
+  private static final Map<String, Binary> binaryCache =
+      new ConcurrentHashMap<>(config.getWORKLOAD_BUFFER_SIZE());
+  private final Session session;
 
-    @Override
-    public void open(boolean enableRPCCompression) throws IoTDBConnectionException {
-      session.open(enableRPCCompression);
-    }
-
-    @Override
-    public void insertRecord(
-        String deviceId,
-        long time,
-        List<String> measurements,
-        List<TSDataType> types,
-        List<Object> values)
-        throws IoTDBConnectionException, StatementExecutionException {
-      session.insertRecord(deviceId, time, measurements, types, values);
-    }
-
-    @Override
-    public void insertAlignedRecord(
-        String multiSeriesId,
-        long time,
-        List<String> multiMeasurementComponents,
-        List<TSDataType> types,
-        List<Object> values)
-        throws IoTDBConnectionException, StatementExecutionException {
-      session.insertAlignedRecord(multiSeriesId, time, multiMeasurementComponents, types, values);
-    }
-
-    @Override
-    public void insertRecords(
-        List<String> deviceIds,
-        List<Long> times,
-        List<List<String>> measurementsList,
-        List<List<TSDataType>> typesList,
-        List<List<Object>> valuesList)
-        throws IoTDBConnectionException, StatementExecutionException {
-      session.insertRecords(deviceIds, times, measurementsList, typesList, valuesList);
-    }
-
-    @Override
-    public void insertAlignedRecords(
-        List<String> multiSeriesIds,
-        List<Long> times,
-        List<List<String>> multiMeasurementComponentsList,
-        List<List<TSDataType>> typesList,
-        List<List<Object>> valuesList)
-        throws IoTDBConnectionException, StatementExecutionException {
-      session.insertAlignedRecords(
-          multiSeriesIds, times, multiMeasurementComponentsList, typesList, valuesList);
-    }
-
-    @Override
-    public void insertTablet(Tablet tablet)
-        throws IoTDBConnectionException, StatementExecutionException {
-      session.insertTablet(tablet);
-    }
-
-    @Override
-    public void insertAlignedTablet(Tablet tablet)
-        throws IoTDBConnectionException, StatementExecutionException {
-      session.insertAlignedTablet(tablet);
-    }
-
-    @Override
-    public void insertRelationalTablet(Tablet tablet)
-        throws IoTDBConnectionException, StatementExecutionException {
-      session.insertRelationalTablet(tablet);
-    }
-
-    @Override
-    public ISessionDataSet executeQueryStatement(String sql)
-        throws IoTDBConnectionException, StatementExecutionException {
-      return new SessionDataSet1(session.executeQueryStatement(sql));
-    }
-
-    @Override
-    public void close() throws IoTDBConnectionException {
-      session.close();
-    }
-
-    @Override
-    public void executeNonQueryStatement(String deleteSeriesSql)
-        throws IoTDBConnectionException, StatementExecutionException {
-      session.executeNonQueryStatement(deleteSeriesSql);
-    }
-
-    private class SessionDataSet1 implements ISessionDataSet {
-      SessionDataSet sessionDataSet;
-
-      public SessionDataSet1(SessionDataSet sessionDataSet) {
-        this.sessionDataSet = sessionDataSet;
-      }
-
-      @Override
-      public RowRecord next() throws IoTDBConnectionException, StatementExecutionException {
-        return sessionDataSet.next();
-      }
-
-      @Override
-      public boolean hasNext() throws IoTDBConnectionException, StatementExecutionException {
-        return sessionDataSet.hasNext();
-      }
-
-      @Override
-      public void close() throws IoTDBConnectionException, StatementExecutionException {
-        sessionDataSet.close();
-      }
-
-      @Override
-      public SessionDataSet.DataIterator iterator() {
-        return sessionDataSet.iterator();
-      }
-    }
-  }
-
-  public IoTDBSession(DBConfig dbConfig) {
+  public IoTDBSession(DBConfig dbConfig) throws IoTDBConnectionException {
     super(dbConfig);
-    LOGGER = LoggerFactory.getLogger(IoTDBSession.class);
     List<String> hostUrls = new ArrayList<>(dbConfig.getHOST().size());
     for (int i = 0; i < dbConfig.getHOST().size(); i++) {
       hostUrls.add(dbConfig.getHOST().get(i) + ":" + dbConfig.getPORT().get(i));
     }
-    sessionWrapper =
-        new BenchmarkSession(
-            new Session.Builder()
-                .nodeUrls(hostUrls)
-                .username(dbConfig.getUSERNAME())
-                .password(dbConfig.getPASSWORD())
-                .enableRedirection(true)
-                .database(dbConfig.getDB_NAME())
-                .version(Version.V_1_0)
-                .sqlDialect(dbConfig.getSQL_DIALECT())
-                .build());
+    if (config.isIoTDB_ENABLE_TABLE())
+      session =
+          new Session.Builder()
+              .nodeUrls(hostUrls)
+              .username(dbConfig.getUSERNAME())
+              .password(dbConfig.getPASSWORD())
+              .enableRedirection(true)
+              .database(dbConfig.getDB_NAME())
+              .version(Version.V_1_0)
+              .sqlDialect(dbConfig.getSQL_DIALECT())
+              .build();
+    else
+      session =
+          new Session.Builder()
+              .nodeUrls(hostUrls)
+              .username(dbConfig.getUSERNAME())
+              .password(dbConfig.getPASSWORD())
+              .enableRedirection(true)
+              .version(Version.V_1_0)
+              .sqlDialect(dbConfig.getSQL_DIALECT())
+              .build();
+  }
+
+  public Status insertOneBatchByTablet(IBatch batch) {
+    Tablet tablet = genTablet(batch);
+    task =
+        service.submit(
+            () -> {
+              try {
+                if (config.isIoTDB_ENABLE_TABLE()) {
+                  session.insertRelationalTablet(tablet);
+                } else if (config.isVECTOR()) {
+                  session.insertAlignedTablet(tablet);
+                } else {
+                  session.insertTablet(tablet);
+                }
+              } catch (IoTDBConnectionException | StatementExecutionException e) {
+                throw new OperationFailException(e);
+              }
+            });
+    return waitWriteTaskToFinishAndGetStatus();
+  }
+
+  public Status insertOneBatchByRecord(IBatch batch) {
+    String deviceId = getDevicePath(batch.getDeviceSchema());
+    int failRecord = 0;
+    List<String> sensors =
+        batch.getDeviceSchema().getSensors().stream()
+            .map(Sensor::getName)
+            .collect(Collectors.toList());
+    for (Record record : batch.getRecords()) {
+      long timestamp = record.getTimestamp();
+      List<TSDataType> dataTypes =
+          constructDataTypes(
+              batch.getDeviceSchema().getSensors(), record.getRecordDataValue().size());
+      List<Object> recordDataValue = convertTypeForBLOB(record, dataTypes);
+      try {
+        if (config.isVECTOR()) {
+          session.insertAlignedRecord(deviceId, timestamp, sensors, dataTypes, recordDataValue);
+        } else {
+          session.insertRecord(deviceId, timestamp, sensors, dataTypes, recordDataValue);
+        }
+      } catch (IoTDBConnectionException | StatementExecutionException e) {
+        failRecord++;
+      }
+    }
+    if (failRecord == 0) {
+      return new Status(true);
+    } else {
+      Exception e = new Exception("failRecord number is " + failRecord);
+      return new Status(false, 0, e, e.toString());
+    }
+  }
+
+  public Status insertOneBatchByRecords(IBatch batch) {
+    List<String> deviceIds = new ArrayList<>();
+    List<Long> times = new ArrayList<>();
+    List<List<String>> measurementsList = new ArrayList<>();
+    List<List<TSDataType>> typesList = new ArrayList<>();
+    List<List<Object>> valuesList = new ArrayList<>();
+    List<String> sensors =
+        batch.getDeviceSchema().getSensors().stream()
+            .map(Sensor::getName)
+            .collect(Collectors.toList());
+    while (true) {
+      String deviceId = getDevicePath(batch.getDeviceSchema());
+      for (Record record : batch.getRecords()) {
+        deviceIds.add(deviceId);
+        times.add(record.getTimestamp());
+        measurementsList.add(sensors);
+        List<TSDataType> dataTypes =
+            constructDataTypes(
+                batch.getDeviceSchema().getSensors(), record.getRecordDataValue().size());
+        valuesList.add(convertTypeForBLOB(record, dataTypes));
+        typesList.add(dataTypes);
+      }
+      if (!batch.hasNext()) {
+        break;
+      }
+      batch.next();
+    }
+    task =
+        service.submit(
+            () -> {
+              try {
+                if (config.isVECTOR()) {
+                  session.insertAlignedRecords(
+                      deviceIds, times, measurementsList, typesList, valuesList);
+                } else {
+                  session.insertRecords(deviceIds, times, measurementsList, typesList, valuesList);
+                }
+              } catch (IoTDBConnectionException | StatementExecutionException e) {
+                throw new OperationFailException(e);
+              }
+            });
+    return waitWriteTaskToFinishAndGetStatus();
+  }
+
+  @Override
+  protected Status executeQueryAndGetStatus(String sql, Operation operation) {
+    String executeSQL;
+    if (config.isIOTDB_USE_DEBUG() && random.nextDouble() < config.getIOTDB_USE_DEBUG_RATIO()) {
+      executeSQL = "debug " + sql;
+    } else {
+      executeSQL = sql;
+    }
+    if (!config.isIS_QUIET_MODE()) {
+      LOGGER.info("{} query SQL: {}", Thread.currentThread().getName(), executeSQL);
+    }
+    AtomicInteger line = new AtomicInteger();
+    AtomicLong queryResultPointNum = new AtomicLong();
+    AtomicBoolean isOk = new AtomicBoolean(true);
+    try {
+      List<List<Object>> records = new ArrayList<>();
+      task =
+          service.submit(
+              () -> {
+                try {
+                  SessionDataSet sessionDataSet = session.executeQueryStatement(executeSQL);
+                  if (config.isIS_COMPARISON()) {
+                    while (sessionDataSet.hasNext()) {
+                      RowRecord rowRecord = sessionDataSet.next();
+                      line.getAndIncrement();
+                      List<Object> record = new ArrayList<>();
+                      switch (operation) {
+                        case AGG_RANGE_QUERY:
+                        case AGG_VALUE_QUERY:
+                        case AGG_RANGE_VALUE_QUERY:
+                          break;
+                        default:
+                          record.add(rowRecord.getTimestamp());
+                          break;
+                      }
+                      List<Field> fields = rowRecord.getFields();
+                      for (int i = 0; i < fields.size(); i++) {
+                        switch (operation) {
+                          case LATEST_POINT_QUERY:
+                            if (i == 0 || i == 2) {
+                              continue;
+                            }
+                          default:
+                            break;
+                        }
+                        record.add(fields.get(i).toString());
+                      }
+                      records.add(record);
+                    }
+                  } else {
+                    SessionDataSet.DataIterator iterator = sessionDataSet.iterator();
+                    while (iterator.next()) {
+                      line.getAndIncrement();
+                    }
+                  }
+
+                  sessionDataSet.close();
+                } catch (StatementExecutionException | IoTDBConnectionException e) {
+                  LOGGER.error("exception occurred when execute query={}", executeSQL, e);
+                  isOk.set(false);
+                }
+                long resultPointNum = line.get();
+                if (!Operation.LATEST_POINT_QUERY.equals(operation)) {
+                  resultPointNum *= config.getQUERY_SENSOR_NUM();
+                  resultPointNum *= config.getQUERY_DEVICE_NUM();
+                }
+                queryResultPointNum.set(resultPointNum);
+              });
+      try {
+        task.get(config.getREAD_OPERATION_TIMEOUT_MS(), TimeUnit.MILLISECONDS);
+      } catch (InterruptedException | ExecutionException | TimeoutException e) {
+        task.cancel(true);
+        return new Status(false, queryResultPointNum.get(), e, executeSQL);
+      }
+      if (isOk.get()) {
+        if (config.isIS_COMPARISON()) {
+          return new Status(true, queryResultPointNum.get(), executeSQL, records);
+        } else {
+          return new Status(true, queryResultPointNum.get());
+        }
+      } else {
+        return new Status(
+            false, queryResultPointNum.get(), new Exception("Failed to execute."), executeSQL);
+      }
+    } catch (Exception e) {
+      return new Status(false, queryResultPointNum.get(), e, executeSQL);
+    } catch (Throwable t) {
+      return new Status(false, queryResultPointNum.get(), new Exception(t), executeSQL);
+    }
+  }
+
+  /**
+   * Using in verification
+   *
+   * @param verificationQuery the query of verification
+   */
+  @Override
+  public Status verificationQuery(VerificationQuery verificationQuery) {
+    DeviceSchema deviceSchema = verificationQuery.getDeviceSchema();
+    List<DeviceSchema> deviceSchemas = new ArrayList<>();
+    deviceSchemas.add(deviceSchema);
+
+    List<Record> records = verificationQuery.getRecords();
+    if (records == null || records.size() == 0) {
+      return new Status(
+          false,
+          new TsdbException("There are no records in verificationQuery."),
+          "There are no records in verificationQuery.");
+    }
+
+    StringBuffer sql = new StringBuffer();
+    sql.append(getSimpleQuerySqlHead(deviceSchemas));
+    Map<Long, List<Object>> recordMap = new HashMap<>();
+    sql.append(" WHERE time = ").append(records.get(0).getTimestamp());
+    recordMap.put(records.get(0).getTimestamp(), records.get(0).getRecordDataValue());
+    for (int i = 1; i < records.size(); i++) {
+      Record record = records.get(i);
+      sql.append(" or time = ").append(record.getTimestamp());
+      recordMap.put(record.getTimestamp(), record.getRecordDataValue());
+    }
+    int point = 0;
+    int line = 0;
+    try {
+      SessionDataSet sessionDataSet = session.executeQueryStatement(sql.toString());
+      while (sessionDataSet.hasNext()) {
+        RowRecord rowRecord = sessionDataSet.next();
+        long timeStamp = rowRecord.getTimestamp();
+        List<Object> values = recordMap.get(timeStamp);
+        for (int i = 0; i < values.size(); i++) {
+          String value = rowRecord.getFields().get(i).toString();
+          String target = String.valueOf(values.get(i));
+          if (!value.equals(target)) {
+            LOGGER.error("Using SQL: " + sql + ",Expected:" + value + " but was: " + target);
+          } else {
+            point++;
+          }
+        }
+        line++;
+      }
+      sessionDataSet.close();
+    } catch (Exception e) {
+      LOGGER.error("Query Error: " + sql);
+      return new Status(false, new TsdbException("Failed to query"), "Failed to query.");
+    }
+    if (recordMap.size() != line) {
+      LOGGER.error(
+          "Using SQL: " + sql + ",Expected line:" + recordMap.size() + " but was: " + line);
+    }
+    return new Status(true, point);
+  }
+
+  @Override
+  public Status deviceQuery(DeviceQuery deviceQuery) {
+    DeviceSchema deviceSchema = deviceQuery.getDeviceSchema();
+    String sql =
+        getDeviceQuerySql(
+            deviceSchema, deviceQuery.getStartTimestamp(), deviceQuery.getEndTimestamp());
+    if (!config.isIS_QUIET_MODE()) {
+      LOGGER.info("IoTDB:" + sql);
+    }
+    List<List<Object>> result = new ArrayList<>();
+    try {
+      SessionDataSet sessionDataSet = session.executeQueryStatement(sql);
+      while (sessionDataSet.hasNext()) {
+        List<Object> line = new ArrayList<>();
+        RowRecord rowRecord = sessionDataSet.next();
+        line.add(rowRecord.getTimestamp());
+        List<Field> fields = rowRecord.getFields();
+        for (Field field : fields) {
+          line.add(field.getStringValue());
+        }
+        result.add(line);
+      }
+      sessionDataSet.close();
+    } catch (Exception e) {
+      LOGGER.error("Query Error: " + sql + " exception:" + e.getMessage());
+      return new Status(false, new TsdbException("Failed to query"), "Failed to query.");
+    }
+
+    return new Status(true, 0, sql, result);
+  }
+
+  @Override
+  public DeviceSummary deviceSummary(DeviceQuery deviceQuery) throws TsdbException {
+    DeviceSchema deviceSchema = deviceQuery.getDeviceSchema();
+    int totalLineNumber;
+    long minTimeStamp, maxTimeStamp;
+    try {
+      SessionDataSet sessionDataSet =
+          session.executeQueryStatement(getTotalLineNumberSql(deviceSchema));
+      RowRecord rowRecord = sessionDataSet.next();
+      totalLineNumber = Integer.parseInt(rowRecord.getFields().get(0).toString());
+      sessionDataSet.close();
+
+      sessionDataSet = session.executeQueryStatement(getMaxTimeStampSql(deviceSchema));
+      rowRecord = sessionDataSet.next();
+      maxTimeStamp = rowRecord.getTimestamp();
+      sessionDataSet.close();
+
+      sessionDataSet = session.executeQueryStatement(getMinTimeStampSql(deviceSchema));
+      rowRecord = sessionDataSet.next();
+      minTimeStamp = rowRecord.getTimestamp();
+      sessionDataSet.close();
+    } catch (IoTDBConnectionException e) {
+      throw new TsdbException("Failed to connect to IoTDB:" + e.getMessage());
+    } catch (StatementExecutionException e) {
+      throw new TsdbException("Failed to execute statement:" + e.getMessage());
+    }
+    return new DeviceSummary(deviceSchema.getDevice(), totalLineNumber, minTimeStamp, maxTimeStamp);
+  }
+
+  protected Tablet genTablet(IBatch batch) {
+    List<IMeasurementSchema> schemaList = new ArrayList<>();
+    List<Tablet.ColumnType> columnTypes = new ArrayList<>();
+    List<Sensor> sensors = batch.getDeviceSchema().getSensors();
+    // All sensors are of type measurement
+    if (config.isIoTDB_ENABLE_TABLE()) {
+      for (int i = 0; i < sensors.size(); i++) {
+        columnTypes.add(Tablet.ColumnType.MEASUREMENT);
+      }
+    }
+    // tag and device as ID column
+    // Add Identity Column Information to Schema
+    sensors.add(new Sensor("device_id", SensorType.STRING));
+    columnTypes.add(Tablet.ColumnType.ID);
+    for (String key : batch.getDeviceSchema().getTags().keySet()) {
+      // Currently, the identity column can only be String
+      sensors.add(new Sensor(key, SensorType.STRING));
+      columnTypes.add(Tablet.ColumnType.ID);
+    }
+    int sensorIndex = 0;
+    for (Sensor sensor : sensors) {
+      SensorType dataSensorType = sensor.getSensorType();
+      schemaList.add(
+          new MeasurementSchema(
+              sensor.getName(),
+              Enum.valueOf(TSDataType.class, dataSensorType.name),
+              Enum.valueOf(TSEncoding.class, getEncodingType(dataSensorType))));
+      sensorIndex++;
+    }
+    // Add the value of the identity column to the value of each record
+    for (int i = 0; i < batch.getRecords().size(); i++) {
+      List<Object> dataValue = batch.getRecords().get(i).getRecordDataValue();
+      dataValue.add(batch.getDeviceSchema().getDevice());
+      for (String key : batch.getDeviceSchema().getTags().keySet()) {
+        dataValue.add(batch.getDeviceSchema().getTags().get(key));
+      }
+    }
+    String deviceId =
+        config.isIoTDB_ENABLE_TABLE()
+            ? batch.getDeviceSchema().getGroup() + "_table"
+            : getDevicePath(batch.getDeviceSchema());
+    Tablet tablet =
+        config.isIoTDB_ENABLE_TABLE()
+            ? new Tablet(deviceId, schemaList, columnTypes, batch.getRecords().size())
+            : new Tablet(deviceId, schemaList, batch.getRecords().size());
+    long[] timestamps = tablet.timestamps;
+    Object[] values = tablet.values;
+
+    for (int recordIndex = 0; recordIndex < batch.getRecords().size(); recordIndex++) {
+      tablet.rowSize++;
+      Record record = batch.getRecords().get(recordIndex);
+      sensorIndex = 0;
+      long currentTime = record.getTimestamp();
+      timestamps[recordIndex] = currentTime;
+      for (int recordValueIndex = 0;
+          recordValueIndex < record.getRecordDataValue().size();
+          recordValueIndex++) {
+        switch (sensors.get(sensorIndex).getSensorType()) {
+          case BOOLEAN:
+            boolean[] sensorsBool = (boolean[]) values[recordValueIndex];
+            sensorsBool[recordIndex] =
+                (boolean) (record.getRecordDataValue().get(recordValueIndex));
+            break;
+          case INT32:
+            int[] sensorsInt = (int[]) values[recordValueIndex];
+            sensorsInt[recordIndex] = (int) (record.getRecordDataValue().get(recordValueIndex));
+            break;
+          case INT64:
+            long[] sensorsLong = (long[]) values[recordValueIndex];
+            sensorsLong[recordIndex] = (long) (record.getRecordDataValue().get(recordValueIndex));
+            break;
+          case FLOAT:
+            float[] sensorsFloat = (float[]) values[recordValueIndex];
+            sensorsFloat[recordIndex] = (float) (record.getRecordDataValue().get(recordValueIndex));
+            break;
+          case DOUBLE:
+            double[] sensorsDouble = (double[]) values[recordValueIndex];
+            sensorsDouble[recordIndex] =
+                (double) (record.getRecordDataValue().get(recordValueIndex));
+            break;
+          case TEXT:
+          case STRING:
+          case BLOB:
+            Binary[] sensorsText = (Binary[]) values[recordValueIndex];
+            sensorsText[recordIndex] =
+                binaryCache.computeIfAbsent(
+                    (String) record.getRecordDataValue().get(recordValueIndex),
+                    BytesUtils::valueOf);
+            break;
+          case TIMESTAMP:
+            long[] sensorsTimestamp = (long[]) values[recordValueIndex];
+            sensorsTimestamp[recordIndex] =
+                (long) (record.getRecordDataValue().get(recordValueIndex));
+            break;
+          case DATE:
+            LocalDate[] sensorsDate = (LocalDate[]) values[recordValueIndex];
+            sensorsDate[recordIndex] =
+                (LocalDate) (record.getRecordDataValue().get(recordValueIndex));
+            break;
+          default:
+            LOGGER.error("Unsupported Type: {}", sensors.get(sensorIndex).getSensorType());
+        }
+        sensorIndex++;
+      }
+    }
+    return tablet;
+  }
+
+  public List<TSDataType> constructDataTypes(List<Sensor> sensors, int recordValueSize) {
+    List<TSDataType> dataTypes = new ArrayList<>();
+    for (int sensorIndex = 0; sensorIndex < recordValueSize; sensorIndex++) {
+      switch (sensors.get(sensorIndex).getSensorType()) {
+        case BOOLEAN:
+          dataTypes.add(TSDataType.BOOLEAN);
+          break;
+        case INT32:
+          dataTypes.add(TSDataType.INT32);
+          break;
+        case INT64:
+          dataTypes.add(TSDataType.INT64);
+          break;
+        case FLOAT:
+          dataTypes.add(TSDataType.FLOAT);
+          break;
+        case DOUBLE:
+          dataTypes.add(TSDataType.DOUBLE);
+          break;
+        case TEXT:
+          dataTypes.add(TSDataType.TEXT);
+          break;
+        case STRING:
+          dataTypes.add(TSDataType.STRING);
+          break;
+        case BLOB:
+          dataTypes.add(TSDataType.BLOB);
+          break;
+        case TIMESTAMP:
+          dataTypes.add(TSDataType.TIMESTAMP);
+          break;
+        case DATE:
+          dataTypes.add(TSDataType.DATE);
+          break;
+      }
+    }
+    return dataTypes;
+  }
+
+  public List<Object> convertTypeForBLOB(Record record, List<TSDataType> dataTypes) {
+    // String change to Binary
+    List<Object> dataValue = record.getRecordDataValue();
+    for (int recordValueIndex = 0;
+        recordValueIndex < record.getRecordDataValue().size();
+        recordValueIndex++) {
+      if (Objects.requireNonNull(dataTypes.get(recordValueIndex)) == TSDataType.BLOB) {
+        dataValue.set(
+            recordValueIndex,
+            binaryCache.computeIfAbsent(
+                (String) record.getRecordDataValue().get(recordValueIndex), BytesUtils::valueOf));
+      }
+    }
+    return dataValue;
+  }
+
+  @Override
+  public Status insertOneBatch(IBatch batch) {
+    DBInsertMode insertMode = dbConfig.getDB_SWITCH().getInsertMode();
+    switch (insertMode) {
+      case INSERT_USE_SESSION_TABLET:
+        return insertOneBatchByTablet(batch);
+      case INSERT_USE_SESSION_RECORD:
+        return insertOneBatchByRecord(batch);
+      case INSERT_USE_SESSION_RECORDS:
+        return insertOneBatchByRecords(batch);
+      default:
+        throw new IllegalStateException("Unexpected INSERT_MODE value: " + insertMode);
+    }
+  }
+
+  Status waitWriteTaskToFinishAndGetStatus() {
+    try {
+      task.get(config.getWRITE_OPERATION_TIMEOUT_MS(), TimeUnit.MILLISECONDS);
+    } catch (InterruptedException | ExecutionException | TimeoutException e) {
+      task.cancel(true);
+      LOGGER.error("insertion failed", e);
+      return new Status(false, 0, e, e.toString());
+    }
+    return new Status(true);
   }
 
   @Override
   public void init() {
     try {
       if (config.isENABLE_THRIFT_COMPRESSION()) {
-        sessionWrapper.open(true);
+        session.open(true);
       } else {
-        sessionWrapper.open();
+        session.open();
       }
       this.service = Executors.newSingleThreadExecutor();
     } catch (IoTDBConnectionException e) {
@@ -201,34 +620,29 @@ public class IoTDBSession extends IoTDBSessionBase {
   public void cleanup() {
     try {
       if (config.isIoTDB_ENABLE_TABLE()) {
-        ISessionDataSet dataSet = sessionWrapper.executeQueryStatement("show databases");
+        SessionDataSet dataSet = session.executeQueryStatement("show databases");
         while (dataSet.hasNext()) {
-          sessionWrapper.executeNonQueryStatement(
+          session.executeNonQueryStatement(
               "drop database " + dataSet.next().getFields().get(0).toString());
         }
       } else {
-        sessionWrapper.executeNonQueryStatement(
+        session.executeNonQueryStatement(
             "drop database root." + config.getDbConfig().getDB_NAME() + ".**");
-        sessionWrapper.executeNonQueryStatement(
-            "drop schema template " + config.getTEMPLATE_NAME());
+        session.executeNonQueryStatement("drop schema template " + config.getTEMPLATE_NAME());
       }
     } catch (IoTDBConnectionException e) {
-      LOGGER.error("Failed to connect to IoTDB:" + e.getMessage());
+      LOGGER.warn("Failed to connect to IoTDB:" + e.getMessage());
     } catch (StatementExecutionException e) {
-      LOGGER.error("Failed to execute statement:" + e.getMessage());
+      LOGGER.warn("Failed to execute statement:" + e.getMessage());
     }
   }
 
   @Override
   public void close() throws TsdbException {
     try {
-      if (sessionWrapper != null) {
-        sessionWrapper.close();
-      }
-      if (ioTDBConnection != null) {
-        ioTDBConnection.close();
-      }
-      this.service.shutdown();
+      if (session != null) session.close();
+      if (ioTDBConnection != null) ioTDBConnection.close();
+      service.shutdown();
     } catch (IoTDBConnectionException ioTDBConnectionException) {
       LOGGER.error("Failed to close session.");
       throw new TsdbException(ioTDBConnectionException);
